{"schema":"decodex/commit/1","summary":"Move frozen edit elements","authority":"manual"}
